SONAR-6370 move the remaining extensions out of core plugin
