+++ server/sonar-server/src/main/java/org/sonar/server/event/package-info.java
@@ -17,5 +17,8 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-@javax.annotation.ParametersAreNonnullByDefault
-package org.sonar.plugins.core.sensors;
+
+@ParametersAreNonnullByDefault
+package org.sonar.server.event;
+
+import javax.annotation.ParametersAreNonnullByDefault;
