+++ server/sonar-server/src/main/java/org/sonar/server/metric/package-info.java
@@ -17,7 +17,8 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
+
 @ParametersAreNonnullByDefault
-package org.sonar.plugins.core.issue;
+package org.sonar.server.metric;
 
 import javax.annotation.ParametersAreNonnullByDefault;
