+++ server/sonar-server/src/test/java/org/sonar/server/event/NewAlertsTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.core.notifications.alerts;
+package org.sonar.server.event;
 
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
