+++ sonar-core/src/main/java/org/sonar/core/resource/DefaultResourceTypes.java
@@ -17,15 +17,15 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.plugins.core;
+package org.sonar.core.resource;
 
-import org.sonar.api.batch.BatchSide;
 import org.sonar.api.ExtensionProvider;
-import org.sonar.api.server.ServerSide;
+import org.sonar.api.batch.BatchSide;
 import org.sonar.api.batch.InstantiationStrategy;
 import org.sonar.api.resources.Qualifiers;
 import org.sonar.api.resources.ResourceType;
 import org.sonar.api.resources.ResourceTypeTree;
+import org.sonar.api.server.ServerSide;
 
 @BatchSide
 @ServerSide
